Allow to reset taglist
The taglist is global and should not be reset, but for unittest it is
usefull to do so. Then it is not a requirement for the tests to use
unique names for tags. The test could accidential fail if a tag is
created that already exists, and it is returned with other values than
the test require.