Refactor Tag to a template
Template to hold the different tag types. They can now be created with
the makeTag<TagType>(args)
In progress to try and get the qt signal to work as it should. It works
with the base class, when initialized. But it can not be put in a
std::variant that is needed to put them in a std::vector.